final css cleanup
